requirements.txt
Django==6.0.4
djangorestframework==3.17.1
django-cors-headers==4.9.0
dj-database-url==2.2.0
python-dotenv==1.0.1
requests==2.33.1
gunicorn==23.0.0
drf-spectacular==0.27.1
psycopg2-binary==2.9.9
whitenoise==6.7.0
